Refactored security tests from SecurityCest Removed get/set di test Removed DestroyTokenCest (covered by CheckTokenCest) Removed GetSessionTokenCest (covered by getRequestToken) Removed GetTokenCest (covered by GetTokenAndKeyCest) Removed GetTokenKeyCest (covered by GetTokenAndKeyCest) Removed GetWorkFactorCest (covered by GetSetWorkFactorCest) Removed GetDefaultHashCest (covered by GetSetDefaultHashCest) Removed SetRandomBytesCest (covered by GetSetRandomBytesCest)
