+++ tests/unit/Security/CheckHashCest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Security;
 
 use UnitTester;
+use Phalcon\Security;
 
 /**
  * Class CheckHashCest
@@ -30,6 +31,34 @@ class CheckHashCest
     public function securityCheckHash(UnitTester $I)
     {
         $I->wantToTest('Security - checkHash()');
-        $I->skipTest('Need implementation');
+        $security = new Security();
+        $password = 'PhalconROCKS!';
+
+        $security->setDefaultHash(Security::CRYPT_DEFAULT);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_STD_DES);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_EXT_DES);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_BLOWFISH);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_BLOWFISH_A);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_BLOWFISH_X);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_BLOWFISH_Y);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_SHA256);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
+
+        $security->setDefaultHash(Security::CRYPT_SHA512);
+        $I->assertTrue($security->checkHash($password, $security->hash($password)));
     }
 }
