+++ tests/unit/Security/CheckTokenCest.php
@@ -13,14 +13,54 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Security;
 
 use UnitTester;
+use Phalcon\Security;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
 
 /**
  * Class CheckTokenCest
  */
 class CheckTokenCest
 {
+    use DiTrait;
+
+    private $shouldStopSession = false;
+
     /**
-     * Tests Phalcon\Security :: checkToken()
+     * executed before each test
+     */
+    public function _before(UnitTester $I)
+    {
+        $I->checkExtensionIsLoaded('openssl');
+
+        $this->setNewFactoryDefault();
+        $this->setDiSessionFiles();
+
+        $_SESSION = [];
+        global $_SESSION;
+    }
+
+    public function _after(UnitTester $I)
+    {
+        if (true === $this->shouldStopSession) {
+            @\session_destroy();
+        }
+    }
+
+    private function startSession(): void
+    {
+        if (PHP_SESSION_ACTIVE !== \session_status()) {
+            @\session_start();
+        }
+
+        if (!isset($_SESSION)) {
+            $_SESSION = [];
+        }
+
+        $this->shouldStopSession = true;
+    }
+
+    /**
+     * Tests Phalcon\Security :: checkToken() and destroyToken()
      *
      * @param UnitTester $I
      *
@@ -30,6 +70,40 @@ class CheckTokenCest
     public function securityCheckToken(UnitTester $I)
     {
         $I->wantToTest('Security - checkToken()');
-        $I->skipTest('Need implementation');
+        $this->startSession();
+
+        $container = $this->getDi();
+        $security  = new Security();
+        $security->setDI($container);
+
+        // Random token and token key check
+        $tokenKey = $security->getTokenKey();
+        $token    = $security->getToken();
+        $_POST    = [$tokenKey => $token];
+        $I->assertTrue($security->checkToken(null, null, false));
+        $I->assertTrue($security->checkToken());
+        $I->assertFalse($security->checkToken());
+
+        // Destroy token check
+        $tokenKey = $security->getTokenKey();
+        $token    = $security->getToken();
+        $security->destroyToken();
+
+        $_POST = [$tokenKey => $token];
+        $I->assertFalse($security->checkToken());
+
+        // Custom token key check
+        $token = $security->getToken();
+        $_POST = ['custom_key' => $token];
+        $I->assertFalse($security->checkToken(null, null, false));
+        $I->assertFalse($security->checkToken('other_custom_key', null, false));
+        $I->assertTrue($security->checkToken('custom_key'));
+
+        // Custom token value check
+        $token = $security->getToken();
+        $_POST = [];
+        $I->assertFalse($security->checkToken(null, null, false));
+        $I->assertFalse($security->checkToken('some_random_key', 'some_random_value', false));
+        $I->assertTrue($security->checkToken('custom_key', $token));
     }
 }
