+++ tests/unit/Security/ConstructCest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Security;
 
 use UnitTester;
+use Phalcon\Security;
 
 /**
  * Class ConstructCest
@@ -30,6 +31,29 @@ class ConstructCest
     public function securityConstruct(UnitTester $I)
     {
         $I->wantToTest('Security - __construct()');
-        $I->skipTest('Need implementation');
+        $security = new Security();
+
+        $class = Security::class;
+        $I->assertInstanceOf($class, $security);
+    }
+
+    /**
+     * Tests the Security constants
+     *
+     * @author Phalcon Team <team@phalconphp.com>
+     * @since  2015-12-19
+     */
+    public function testSecurityConstants(UnitTester $I)
+    {
+        $I->assertEquals(0, Security::CRYPT_DEFAULT);
+        $I->assertEquals(1, Security::CRYPT_STD_DES);
+        $I->assertEquals(2, Security::CRYPT_EXT_DES);
+        $I->assertEquals(3, Security::CRYPT_MD5);
+        $I->assertEquals(4, Security::CRYPT_BLOWFISH);
+        $I->assertEquals(5, Security::CRYPT_BLOWFISH_A);
+        $I->assertEquals(6, Security::CRYPT_BLOWFISH_X);
+        $I->assertEquals(7, Security::CRYPT_BLOWFISH_Y);
+        $I->assertEquals(8, Security::CRYPT_SHA256);
+        $I->assertEquals(9, Security::CRYPT_SHA512);
     }
 }
